+++ R/app_server.R
@@ -38,6 +38,8 @@
 #' @noRd
 
 app_server <- function(input, output, session)  {
+ down_color = "#6a73cf"
+    up_color ="#f26115"
 mySetTheme <- function()
   {
     mySetTheme <- theme_prism(base_size = 14, border = TRUE) +
@@ -66,7 +68,7 @@ mySetTheme <- function()
   } else {
     tree <- tree()
   }
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none")
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none")
     x <- as.numeric(input$plotClick$x)
     y <- as.numeric(input$plotClick$y)
     node <- click_node(x, y, p$data)
@@ -87,6 +89,7 @@ mySetTheme <- function()
  )
   sub_tree <- eventReactive(
     input$node, {
+#      browser()
       if(input$node != "") {
         tree <- tree()
         check_node <- as.numeric(input$node)<length(as.phylo(tree)$tip.label)
@@ -108,11 +111,11 @@ mySetTheme <- function()
     req(!is.null(input$treefile))
     if (input$filetype=="Newick") {
       tree <- read.tree(input$treefile$datapath) %>% as.phylo()
-    } else if (input$filetype=="beast") {
+    } else if (input$filetype=="Beast") {
       tree <- read.beast(input$treefile$datapath) %>% as.phylo()
-    } else if (input$filetype=="Nexus") {
+    } else if (input$filetype=="NEXUS") {
       tree <- read.nexus(input$treefile$datapath)
-    } else if (input$filetype=="phylip") {
+    } else if (input$filetype=="Phylip") {
       tree <- read.phylip.tree(input$treefile$datapath) %>% as.phylo()
     }
     tree
@@ -210,6 +213,47 @@ label_all<-reactive({
   updateTextInput(session,"node",value = reset_root_node)
 })
 
+   output$Summary <- renderTable({  
+      date <- date()|> unlist()|>as.numeric()
+      divergence <- divergence()|>unlist()|>as.numeric()
+      df <- cbind(divergence, date)|>as.data.frame()
+      if (input$format == "yy" | input$format == "yyyy") {
+        range <- max(date) - min(date)
+      }
+      else{
+        range <- max(date) - min(date)
+        range <- range * 365
+      }
+      ##make a summary and output
+      summary <- data.frame(Dated.tips=c("Date range", "Slope(rate)", 
+                                         "X-Intercept", "Correlation", 
+                                         "R squared", "RSE"), value=numeric(6))
+      summary[1, 2] <- range
+      summary[4, 2] <- as.numeric(cor(date, divergence))
+      lm.rtt <- lm(df)
+      summary[2, 2] <- as.numeric(lm.rtt$coefficients[2])
+      summary[3, 2] <- as.numeric(
+        abs(lm.rtt$coefficients[1])) / as.numeric(lm.rtt$coefficients[2])
+      summary[5, 2] <- summary(lm.rtt)$r.squared
+      summary[6, 2] <- summary(lm.rtt)[["sigma"]]
+      # summary[7, 2] <- shapiro.test(rstudent(lm(df)))[2]
+      # summary[8, 2] <- DescTools::RunsTest(rstudent(lm(df)))$p.value
+      table_dt(summary)
+      print(summary)
+    },
+    digits = 5, width = 8)
+    
+    table_dt <- reactiveVal()
+    output$download_dt2 <- downloadHandler(
+      filename = function(){
+        "summary(reset).csv"
+      },
+      content = function(file){
+        df <- table_dt()
+        write.csv(df,file,row.names = FALSE)
+      }
+    )
+
 observeEvent(input$reset2,{
     output$Summary2 <- renderTable({  
       df1 <- merged_data()
@@ -227,6 +271,8 @@ observeEvent(input$reset2,{
        print(summary)
     })
   output$plot1 <- renderPlot({
+    # browser()
+    req(sub_tree())
   if (input$choose_analysis=="Temporal_signal") {
     tree <- sub_tree()
     if (is.null(tree)) {
@@ -237,18 +283,18 @@ observeEvent(input$reset2,{
     down.table <- down.table()
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -265,18 +311,18 @@ observeEvent(input$reset2,{
     
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -287,7 +333,7 @@ observeEvent(input$reset2,{
       return(NULL)
     }
 
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {  
       p <- p+ geom_tiplab(size=input$tipsize)
@@ -335,10 +381,10 @@ observeEvent(input$reset2,{
     p <- p_all2+
       geom_point(data=df, aes_string(x = input$x_var, y =input$y_var)) +
       geom_smooth(data=df, aes_string(x = input$x_var, y =input$y_var),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = 'blue') +
-      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color ="red")+
+      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = down_color) +
+      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() +
+      mySetTheme()+theme_classic() +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
     plotd3(p)
     print(p)
@@ -350,10 +396,10 @@ observeEvent(input$reset2,{
     p <- ggplot(df, aes_string(x = input$x_var, y =input$y_var)) +
       geom_point() +
       geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = 'blue') +
-      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color ="red")+
+      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = down_color) +
+      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() +
+      mySetTheme()+theme_classic() +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
     plotd3(p)
     print(p)
@@ -415,7 +461,7 @@ observeEvent(input$reset2,{
   
   output$plot1 <- renderPlot({
     # browser()
-    req(sub_tree())
+    # req(sub_tree())
   if (input$choose_analysis=="Temporal_signal") {
     tree <- sub_tree()
     if (is.null(tree)) {
@@ -426,18 +472,18 @@ observeEvent(input$reset2,{
     down.table <- down.table()
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -454,18 +500,18 @@ observeEvent(input$reset2,{
     
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -476,7 +522,7 @@ observeEvent(input$reset2,{
       return(NULL)
     }
 
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {  
       p <- p+ geom_tiplab(size=input$tipsize)
@@ -507,6 +553,7 @@ observeEvent(input$reset2,{
   
   #2.取出日期
   output$datetable <- renderDataTable({
+    # browser()
      tree <- sub_tree()
     if (is.null(tree)) {
       return()
@@ -558,10 +605,10 @@ observeEvent(input$reset2,{
       geom_point(data=df, aes(x = date, y = divergence+sub_divergence)) +
       geom_smooth(data=df, aes(x = date, y = divergence+sub_divergence),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)+
-      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = 'blue') +
-      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color ="red")+
+      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = down_color) +
+      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme()
+      mySetTheme()+theme_classic()
       plotd2(p)
       p
   }else {
@@ -572,10 +619,10 @@ observeEvent(input$reset2,{
       geom_point() +
       geom_smooth(data=df, aes(x = date, y = divergence+sub_divergence),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)+
-      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = 'blue') +
-      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color ="red")+
+      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = down_color) +
+      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() 
+      mySetTheme()+theme_classic() 
       plotd2(p)
       p
   }
@@ -591,10 +638,10 @@ observeEvent(input$reset2,{
   #   p <- ggplot(df, aes(x = date, y = divergence)) +
   #     geom_point() +
   #     geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-  #     geom_point(data = down.table, aes(x = date, y = divergence), color = 'blue') +
-  #     geom_point(data = up.table,aes(x = date, y = divergence), color ="red")+
+  #     geom_point(data = down.table, aes(x = date, y = divergence), color = down_color) +
+  #     geom_point(data = up.table,aes(x = date, y = divergence), color =up_color)+
   #     # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-  #     mySetTheme() +
+  #     mySetTheme()+theme_classic() +
   #     stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
   #   plotd2(p)
   #   print(p)
@@ -676,7 +723,7 @@ observeEvent(input$reset2,{
              stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)+
               geom_point(data = exclude.table, aes(x = date, y = divergence+sub_divergence), color = 'gray')+
              # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-             mySetTheme()
+             mySetTheme()+theme_classic()
              plotd2(p)
              p
          }else {
@@ -685,7 +732,7 @@ observeEvent(input$reset2,{
         geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
         geom_point(data = exclude.table, aes(x = date, y = divergence+sub_divergence), color = 'gray') +
         # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-        mySetTheme() +
+        mySetTheme()+theme_classic() +
         stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
       plotd2(p)
       print(p)
@@ -699,7 +746,7 @@ observeEvent(input$reset2,{
       #   geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
       #   geom_point(data = exclude.table, aes(x = date, y = divergence), color = 'gray') +
       #   # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      #   mySetTheme() +
+      #   mySetTheme()+theme_classic() +
       #   stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
       # plotd2(p)
       # print(p)
@@ -758,7 +805,7 @@ observeEvent(input$reset2,{
       to_drop <- c(down.table$label,up.table$label)
       tip_reduced <- drop.tip(tree, to_drop)
 
-      p <- ggtree(tip_reduced,color=input$color3, size=input$size)+mySetTheme2() 
+      p <- ggtree(tip_reduced,color=input$color3, size=input$size) 
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       if (input$tip) {
         p <- p+ geom_tiplab(size=input$tipsize)
@@ -821,7 +868,7 @@ output$plot3 <- renderPlot({
         geom_smooth(data=keep.table, aes_string(x = input$x_var, y = input$y_var),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
         geom_point(data = need.exclude.table, aes_string(x = input$x_var, y = input$y_var), color = 'gray') +
         # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-        mySetTheme() +
+        mySetTheme()+theme_classic() +
         stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
       plotd3(p)
       print(p)
@@ -833,7 +880,7 @@ output$plot3 <- renderPlot({
         geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
         geom_point(data = need.exclude.table, aes_string(x = input$x_var, y = input$y_var), color = 'gray') +
         # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-        mySetTheme() +
+        mySetTheme()+theme_classic() +
         stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
       plotd3(p)
       print(p)
@@ -905,18 +952,18 @@ output$plot1 <- renderPlot({
     down.table <- down.table()
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -933,18 +980,18 @@ output$plot1 <- renderPlot({
     
     d <- rbind(up.table,down.table)
     all <- merge(d,df,by='label',all = T)
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {
       
       p <- p+ geom_tiplab(size=input$tipsize)
       p <- p%<+%all+geom_tiplab(aes(color=category))+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }else if(input$tip_point){
       if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
       p <- p%<+%all+ geom_tippoint(aes(color=category),size=input$tipsize)+
         geom_tippoint(aes(color=category),size=input$tipsize)+
-        scale_color_manual(values = c("up" = "red", "down" = "blue"))
+        scale_color_manual(values = c("up" = up_color, "down" = down_color))
     }
     plotd1(p)
     p
@@ -955,7 +1002,7 @@ output$plot1 <- renderPlot({
       return(NULL)
     }
 
-    p <- ggtree(tree,color=input$color3, size=input$size)+mySetTheme2()+theme(legend.position="none") 
+    p <- ggtree(tree,color=input$color3, size=input$size)+theme(legend.position="none") 
     if(input$geom_nodelab){p <- p+geom_nodelab(aes(label=node),hjust=-.3)}
     if (input$tip) {  
       p <- p+ geom_tiplab(size=input$tipsize)
@@ -995,10 +1042,10 @@ output$plot2 <- renderPlot({
       geom_point(data=df, aes(x = date, y = divergence+sub_divergence)) +
       geom_smooth(data=df, aes(x = date, y = divergence+sub_divergence),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)+
-      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = 'blue') +
-      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color ="red")+
+      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = down_color) +
+      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme()
+      mySetTheme()+theme_classic()
       plotd2(p)
       p
   }else {
@@ -1009,10 +1056,10 @@ output$plot2 <- renderPlot({
       geom_point() +
       geom_smooth(data=df, aes(x = date, y = divergence+sub_divergence),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)+
-      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = 'blue') +
-      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color ="red")+
+      geom_point(data = down.table, aes(x = date, y = divergence+sub_divergence), color = down_color) +
+      geom_point(data = up.table,aes(x = date, y = divergence+sub_divergence), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() 
+      mySetTheme()+theme_classic() 
       plotd2(p)
       p
   }
@@ -1028,10 +1075,10 @@ output$plot2 <- renderPlot({
   #   p <- ggplot(df, aes(x = date, y = divergence)) +
   #     geom_point() +
   #     geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-  #     geom_point(data = down.table, aes(x = date, y = divergence), color = 'blue') +
-  #     geom_point(data = up.table,aes(x = date, y = divergence), color ="red")+
+  #     geom_point(data = down.table, aes(x = date, y = divergence), color = down_color) +
+  #     geom_point(data = up.table,aes(x = date, y = divergence), color =up_color)+
   #     # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-  #     mySetTheme() +
+  #     mySetTheme()+theme_classic() +
   #     stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
   #   plotd2(p)
   #   print(p)
@@ -1248,10 +1295,10 @@ df1 <- merged_data()
     p <- ggplot(df, aes_string(x = input$x_var, y =input$y_var)) +
       geom_point() +
       geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = 'blue') +
-      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color ="red")+
+      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = down_color) +
+      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() +
+      mySetTheme()+theme_classic() +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
     plotd3(p)
 
@@ -1326,10 +1373,10 @@ df1 <- merged_data()
     p <- p_all2+
       geom_point(data=df, aes_string(x = input$x_var, y =input$y_var)) +
       geom_smooth(data=df, aes_string(x = input$x_var, y =input$y_var),method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = 'blue') +
-      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color ="red")+
+      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = down_color) +
+      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() +
+      mySetTheme()+theme_classic() +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
     plotd3(p)
     print(p)
@@ -1341,10 +1388,10 @@ df1 <- merged_data()
     p <- ggplot(df, aes_string(x = input$x_var, y =input$y_var)) +
       geom_point() +
       geom_smooth(method = "lm", se = FALSE, formula = y ~ x,colour=input$color2) +
-      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = 'blue') +
-      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color ="red")+
+      geom_point(data = need.down.table, aes_string(x = input$x_var, y =input$y_var), color = down_color) +
+      geom_point(data = need.up.table,aes_string(x = input$x_var, y =input$y_var), color =up_color)+
       # geom_text(data = d, aes(x = date, y = divergence, label = label)) +
-      mySetTheme() +
+      mySetTheme()+theme_classic() +
       stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), parse = TRUE)
     plotd3(p)
     print(p)
@@ -1444,6 +1491,7 @@ df1 <- merged_data()
   )
     
     output$Summary <- renderTable({  
+      # browser()
       date <- date()|> unlist()|>as.numeric()
       divergence <- divergence()|>unlist()|>as.numeric()
       df <- cbind(divergence, date)|>as.data.frame()