+++ inst/doc/shinyTempSignal.R
@@ -1,0 +1,12 @@
+## ----include=FALSE------------------------------------------------------------
+knitr::opts_chunk$set(warning = FALSE,
+                      message = TRUE)
+library(yulab.utils)                  
+
+## ----eval=FALSE---------------------------------------------------------------
+#  install.packages("shinyTempSignal")
+
+## ----eval=FALSE---------------------------------------------------------------
+#  library(shinyTempSignal)
+#  run_shinyTempSignal()
+