+++ NEWS.md
@@ -1,3 +1,9 @@
+# shinyTempSignal 0.0.6
+
++ add vignette (2024-01-20, Sat)
++ modify color scheme (2024-01-15, Mon, #30)
++ update output summary data frame (2024-01-13, Sat, #29)
+
 # shinyTempSignal 0.0.5
 
 + simultaneous update node across different panels (2023-12-12, Tue, #28)